+++ NEWS.md
@@ -8,6 +8,8 @@ Major news relating to breaking changes in ApproxManifoldProducts.jl
 - Use `HomotopyDensity_legacy(; partial)` instead of `HomotopyDensity{partial}(.;.)` (breaking change)
 - Refactored `HomotopyDensity` parameters. (breaking change)
 - Rename `HomotopyDensity.observability` instead of `.infoPerCoord` (breaking change)
+- Rename `HomotopyDensity.elements` instead of `.data` (breaking change)
+- Rename `Homotopy.structure` instead of `.geometric_permute` (breaking change)
 
 ## v0.14 (26Q2)